geckoapi1/pkg/connectors/coingecko/coingecko.go
package coingecko

import (
	"fmt"
	"log"
	"strings"

	gecko "github.com/superoo7/go-gecko/v3"
)

type GeckoApi struct {
	URL string
}

func NewGeckoApi() *GeckoApi {
	return &GeckoApi{
		URL: "https://api.coingecko.com/api/v3",
	}
}

var cg = gecko.NewClient(nil)

func (g *GeckoApi) LoadCoins() (int, error) {
	list, err := cg.CoinsList()
	if err != nil {
		log.Fatal(err)
	}
	fmt.Println("Available coins:")

	return len(*list), nil
}

func (g *GeckoApi) GetRates(from, to string) (map[string]interface{}, error) {
	var ticker = map[string]string{
		"BTC":  "bitcoin",
		"ETH":  "ethereum",
		"USDT": "usd",
	}
	fromParam := []string{strings.ToLower(ticker[from])}
	toParam := []string{strings.ToLower(ticker[to])}
	result := make(map[string]interface{})

	sp, err := cg.SimplePrice(fromParam, toParam)
	if err != nil {
		return nil, fmt.Errorf("failed to fetch rates: %w", err)
	}

	if rate, ok := (*sp)[from][to]; ok {
		result[to] = float64(rate)
		return result, nil
	} else {
		return nil, fmt.Errorf("conversion rate not found")
	}
}
